Revert "Code cleanup"
