+++ src/components/CreateIssue/CreateForm.tsx
@@ -9,7 +9,7 @@ import { Issue } from "@/lib/types";
 import { verifyUrlParam } from "@/lib/utils";
 import FormTextField from "../Form/FormTextField";
 import TextFieldLabel from "../Form/TextFieldLabel";
-import { EMPTY_FORM } from "./FormConstants";
+import { EMPTY_FORM, FormValues } from "./FormConstants";
 import StatusSelect from "./StatusSelect";
 import PrioritySelect from "@/components/CreateIssue/PrioritySelect";
 import { useSetAtom } from "jotai";
@@ -17,7 +17,7 @@ import { addOneIssueAtom } from "@/components/utils/atom";
 import { createGridRowFromIssue } from "@/components/Backlog/utils";
 import AppError from "@/lib/service/AppError";
 import { LoadingButton } from "@mui/lab";
-import { IssueFormValues } from "@/lib/types/Issue";
+import { IssueFormValues } from "@/lib/service/Issue/IssueRequest";
 
 const FormRow = styled(Grid)(() => ({
   "&.MuiGrid-item": {
