+++ src/lib/auth/UserAuth.ts
@@ -1,4 +1,4 @@
-import userRepo from "@/lib/dao/UserRepository";
+import userRepo from "@/server/domain/UserRepository";
 import { User } from "@/lib/types";
 import bcrypt from "bcrypt";
 import { NewUser } from "@/lib/types/User";
