+++ src/lib/client-data/issues.ts
@@ -1,6 +1,7 @@
 import { Issue, PriorityType, StatusType } from "lib/types";
 import AppError from "@/lib/service/AppError";
-import { IssueData, IssueFormValues } from "@/lib/types/Issue";
+import { IssueFormValues } from "@/lib/service/Issue/IssueRequest";
+import { IssueData } from "@/lib/types/Issue";
 
 export function toStatusString(
   status: number | undefined,
