+++ src/lib/service/Issue/IssueRequest.ts
@@ -1,48 +1,89 @@
-import { StatusType } from "@/lib/types";
-import StatusRepository from "@/lib/dao/StatusRepository";
+import { PriorityType, StatusType } from "@/lib/types";
+import StatusRepository from "@/server/domain/StatusRepository";
 import { LexoRank } from "lexorank";
-import { serverOnly } from "@/lib/utils";
-import { IssueFormValues } from "@/lib/types/Issue";
+import Issue from "./index";
 
-export default class IssueRequest {
-  private _id?: number = undefined;
-  private readonly values: IssueFormValues = {};
+export interface IssueFormValues {
+  title?: string;
+  description?: string;
+  status?: number;
+  priority?: number;
+}
+
+export default class IssueRequest implements IssueFormValues {
+  public id?: number = undefined;
+  public title?: string = undefined;
+  public description?: string = undefined;
+  public status?: number;
+  public priority?: number;
+  public assignee?: string;
+  public key?: string;
   private _order?: LexoRank;
-  private _statusDb: StatusRepository | null = null;
-  private _key?: string = undefined;
-
-  public constructor(value: IssueFormValues) {
-    serverOnly();
-    this.values = { ...value };
-  }
-
-  public get id(): number | undefined {
-    return this._id;
-  }
-  public get order(): string {
-    return this._order?.toString() || "";
-  }
-  public get priority(): number | undefined {
-    return this.values.priority;
-  }
-  public get status(): number | undefined {
-    return this.values.status;
-  }
-  public get title(): string | undefined {
-    return this.values.title;
-  }
+  private _validStatuses?: StatusType[];
+  private _validPriorities?: PriorityType[];
 
   public set order(value: string) {
     this._order = LexoRank.parse(value);
   }
-  public set key(value: string) {
-    if (value.length > 5) return;
-    this._key = value;
+  public set validStatuses(value: StatusType[]) {
+    this._validStatuses = value;
+  }
+  public set validPriorities(value: PriorityType[]) {
+    this._validPriorities = value;
   }
 
-  public getEncodedDescription(): Buffer | undefined {
-    if (!this.validDescription()) return undefined;
-    return Buffer.from(this.values.description!);
+  public constructor(value: IssueFormValues) {
+    this.title = value.title;
+    this.description = value.description;
+    this.status = value.status;
+    this.priority = value.priority;
+  }
+
+  public get order(): string {
+    return this._order?.toString() || "";
+  }
+
+  public createIssue(): Issue | undefined {
+    let newIssue: Issue | undefined = undefined;
+    if (!this._validPriorities || !this._validStatuses) {
+      return undefined;
+    }
+
+    // Find issue and priority type
+    const foundPriority = this._validPriorities.find(
+      priority => priority.id === this.priority
+    );
+    const foundStatus = this._validStatuses.find(
+      status => status.id === this.status
+    );
+
+    if (
+      this.id &&
+      this.title &&
+      foundStatus &&
+      foundPriority &&
+      this.key &&
+      this._order
+    ) {
+      newIssue = new Issue(
+        this.id,
+        this.title,
+        foundStatus,
+        foundPriority,
+        this.key,
+        this._order
+      );
+    }
+
+    if (newIssue) {
+      if (this.assignee) {
+        newIssue.assignee = this.assignee;
+      }
+      if (this.description) {
+        newIssue.description = this.description;
+      }
+    }
+    return newIssue;
   }
 
   public saveStatus(value: number): void {
@@ -50,7 +91,7 @@ export default class IssueRequest {
       return;
     }
 
-    this.values.status = value;
+    this.status = value;
   }
 
   public async verifyEntries(db: StatusRepository): Promise<boolean> {
@@ -58,13 +99,17 @@ export default class IssueRequest {
   }
 
   private async verifyStatus(db: StatusRepository): Promise<boolean> {
-    this._statusDb = db;
-    if (!this.isNumber(this.status)) return false;
+    if (Number.isNaN(this.status) || this.status === undefined) {
+      console.log("Status is not a number.");
+      return false;
+    }
 
     // Status provided is not defined in database
-    const validStatuses: number[] = await this.getValidStatuses();
-    if (!validStatuses.includes(Number(this.status))) {
-      console.error("Invalid status");
+    const validStatuses: number[] = (await db.fetchStatuses()).map(
+      (value: StatusType): number => value.id
+    );
+    if (validStatuses.indexOf(this.status) === -1) {
+      console.log("Invalid status");
       return false;
     }
 
@@ -72,37 +117,14 @@ export default class IssueRequest {
   }
 
   private verifyTitle(): boolean {
-    return !(this.title === undefined || this.title.length === 0);
-  }
-
-  private isNumber(value?: number): boolean {
-    if (value === undefined || Number.isNaN(value)) {
-      console.error("Value is not a number");
+    if (this.title === undefined) {
       return false;
     }
-    return true;
-  }
 
-  private async getValidStatuses(): Promise<number[]> {
-    if (!this.statusDbExists()) return [];
-
-    return (await this._statusDb!.fetchStatuses()).map(
-      (value: StatusType): number => value.id
-    );
-  }
-
-  private statusDbExists(): boolean {
-    if (this._statusDb === null) {
-      console.error("Status repository not set");
+    if (this.title.length === 0) {
       return false;
     }
-    return true;
-  }
 
-  private validDescription(): boolean {
-    return (
-      this.values.description !== undefined &&
-      this.values.description.length > 0
-    );
+    return true;
   }
 }
