+++ src/lib/service/Project.ts
@@ -1,8 +1,8 @@
-import IssueRepository from "@/lib/dao/IssueRepository";
+import IssueRepository from "@/server/domain/IssueRepository";
 import IssueRequest from "@/lib/service/Issue/IssueRequest";
 import { Issue } from "@/lib/types";
-import { IIssueDB } from "@/lib/dao/interfaces";
-import StatusRepository from "@/lib/dao/StatusRepository";
+import { IIssueDB } from "@/server/domain/interfaces";
+import StatusRepository from "@/server/domain/StatusRepository";
 
 /**
  * Project is a service class that handles the business logic for the project.
@@ -39,9 +39,9 @@ export default class Project {
       if (issue.key !== undefined) {
         issue.id = this.getIssueId(issue.key);
       }
-      return this._store.editExistingIssue(issue);
+      return this._store.editIssue(issue);
     }
-    return this._store.saveNewIssue(issue);
+    return this._store.saveIssue(issue);
   }
 
   public async getAllIssues(): Promise<Issue[]> {
@@ -51,9 +51,7 @@ export default class Project {
   public async getIssue(key: string): Promise<Issue> {
     const issueId: number = this.getIssueId(key);
 
-    const foundIssue: Issue | null = await this._store.fetchOneIssueWithId(
-      issueId
-    );
+    const foundIssue: Issue | null = await this._store.fetchIssue(issueId);
 
     if (foundIssue === null) {
       throw new Error("Issue not found");
