+++ src/pages/api/projects.ts
@@ -1,9 +1,9 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { Project } from "lib/types";
-import ProjectRepository from "@/lib/dao/ProjectRepository";
+import ProjectRepository from "@/server/domain/ProjectRepository";
 import { JWT } from "next-auth/jwt";
 import { getUserToken } from "@/lib/auth/session";
-import { IProjectDB } from "@/lib/dao/interfaces";
+import { IProjectDB } from "@/server/domain/interfaces";
 import ProjectRequest from "@/lib/service/ProjectRequest";
 import NextjsProjectRequest from "@/lib/service/NextjsProjectRequest";
 import AppError from "@/lib/service/AppError";
