+++ src/server/domain/ProjectRepository.ts
@@ -1,6 +1,6 @@
 import { Prisma } from "@prisma/client";
-import { prisma } from "@/lib/dao/prisma";
-import { Project } from "@/lib/types";
+import { prisma } from "@/server/domain/prisma";
+import { Project } from "lib/types";
 import { IProjectDB } from "./interfaces";
 import ProjectRequest from "@/lib/service/ProjectRequest";
