+++ src/server/domain/StatusRepository.ts
@@ -1,5 +1,5 @@
 import { StatusType } from "@/lib/types";
-import { prisma } from "@/lib/dao/prisma";
+import { prisma } from "@/server/domain/prisma";
 import { Prisma } from "@prisma/client";
 
 export default class StatusRepository {
