+++ src/server/domain/UserRepository.ts
@@ -1,7 +1,7 @@
 import { Prisma } from "@prisma/client";
-import { prisma } from "@/lib/dao/prisma";
+import { prisma } from "@/server/domain/prisma";
 import { NewUser, User } from "@/lib/types/User";
-import { IUserDB } from "@/lib/dao/interfaces";
+import { IUserDB } from "@/server/domain/interfaces";
 import { USER_CREATION_ERROR } from "@/lib/client-data/errors";
 import AppError from "@/lib/service/AppError";
