+++ src/server/domain/interfaces.ts
@@ -10,10 +10,10 @@ export interface IProjectDB {
 
 export interface IIssueDB {
   fetchAllIssues(): Promise<Issue[]>;
-  saveNewIssue(req: IssueRequest): Promise<Issue>;
-  editExistingIssue(req: IssueRequest): Promise<Issue>;
+  saveIssue(req: IssueRequest): Promise<Issue>;
+  editIssue(req: IssueRequest): Promise<Issue>;
   deleteIssue(id: number): Promise<void>;
-  fetchOneIssueWithId(id: number): Promise<Issue | null>;
+  fetchIssue(id: number): Promise<Issue | null>;
 }
 
 export interface IUserDB {
